[FEAT] 지도 관련 기능 추가

See merge request s09-webmobile4-sub2/S09P12D104!11

+++ frontend/android/app/src/main/java/com/ssafy/hifes/data/model/OrganizedFestivalDto.kt
@@ -11,6 +11,6 @@ data class OrganizedFestivalDto(
     var fesEndDate: Date,
     var fesOutline: String,
     var fesAddress: String,
-    var fesLatitude: Float,
-    var fesLongitude: Float
+    var fesLatitude: Double,
+    var fesLongitude: Double
 )
